// apps/travelling-api/src/libs/types/common.ts
import { registerEnumType } from "@nestjs/graphql";
import { ObjectId } from "mongoose";

export enum Message {
  SOMETHING_WENT_WRONG = "Something went wrong!",
  NO_DATA_FOUND = "No data found!",
  CREATE_FAILED = "Create failed!",
  UPDATE_FAILED = "Update failed!",
  REMOVE_FAILED = "Remove failed!",
  UPLOAD_FAILED = "Upload failed!",
  BAD_REQUEST = "Bad Request",

  USED_MEMBERNICK_OR_PHONE="Used memberNick or phone",

  NO_MEMBER_NICK = "No member with that member nick!",
  BLOCKED_USER = "You have been blocked!",
  WRONG_PASSWORD = "Wrong password, try again!",
  NOT_AUTHENTICATED = "You are not authenticated, please login first!",
  TOKEN_NOT_EXIST = "Bearer Token is not provided!",
  ONLY_SPECIFIC_ROLES_ALLOWED = "Allowed only for members with specific roles!",
  NOT_ALLOWED_REQUEST = "Not allowed request!",
  PROVIDE_ALLOWED_PHOTO = "Please provide jpg, jpeg or png images!",
  SELF_SUBSCRIPTION_DENIED = "Self subscription is denied!",
}
export enum Direction{ //agentlarni pasdan yuqorig yoki yuqoridan pasga qarab tartiblash uchun enum
    ASC =1,
    DESC =-1,
}

export interface T {
    [key: string]: any
}

export interface StatisticModifier {
    _id: ObjectId;     //kerakli collectionning document idsi 
    targetKey: string; //kerakli qiymatni ozgartiramiz.yani nimani
    modifier: number;    //qanday qiymatga ozgartirmoqchimiz?
}


registerEnumType(Direction,{name: 'Direction'})